Merge branch 'develop' of https://framework.allotrope.org/osthus-extensions/osthus-ambeth into develop

Conflicts:
	jambeth/pom.xml

+++ jambeth/jambeth-query/src/main/java/de/osthus/ambeth/persistence/IVersionItem.java
@@ -6,9 +6,9 @@ public interface IVersionItem extends IDisposable
 {
 	Object getId();
 
-	Object getId(byte idIndex);
+	Object getId(int idIndex);
 
 	Object getVersion();
 
-	byte getAlternateIdCount();
+	int getAlternateIdCount();
 }
